requirements.txt
Flask==3.0.2
opencv-python==4.9.0.80
numpy==1.26.4
google-generativeai==0.4.1
python-dotenv==1.0.1
websocket-client==1.7.0
requests==2.31.0
PyMuPDF==1.24.0
easyocr==1.7.2
Werkzeug==3.0.1
pyttsx3==2.90
SpeechRecognition==3.14.0
